+++ skbio/diversity/beta/_base.py
@@ -18,20 +18,20 @@ from skbio.util._decorator import experimental, deprecated
 
 
 @experimental(as_of="0.4.0")
-def pw_distances(counts, ids=None, metric="braycurtis", **kwargs):
+def pw_distances(metric, counts, ids=None, **kwargs):
     """Compute distances between all pairs of columns in a counts matrix
 
     Parameters
     ----------
+    metric : str, callable
+        The pairwise distance function as a string or callable to use when
+        generating pairwise distances. See the scipy ``pdist`` docs and the
+        scikit-bio functions linked under *See Also* for available metrics.
     counts : 2D array_like of ints or floats
         Matrix containing count/abundance data where each row contains counts
         of observations in a given sample.
     ids : iterable of strs, optional
         Identifiers for each sample in ``counts``.
-    metric : str, optional
-        The name of the pairwise distance function to use when generating
-        pairwise distances. See the scipy ``pdist`` docs, linked under *See
-        Also*, for available metrics.
 
     Returns
     -------
@@ -46,6 +46,8 @@ def pw_distances(counts, ids=None, metric="braycurtis", **kwargs):
 
     See Also
     --------
+    unweighted_unifrac
+    weighted_unifrac
     scipy.spatial.distance.pdist
     pw_distances_from_table
 
@@ -70,18 +72,18 @@ pw_distances_from_table_deprecation_reason = (
 
 @deprecated(as_of="0.4.0", until="0.4.1",
             reason=pw_distances_from_table_deprecation_reason)
-def pw_distances_from_table(table, metric="braycurtis"):
+def pw_distances_from_table(metric, table):
     """Compute distances between all pairs of samples in table
 
     Parameters
     ----------
+    metric : str, callable
+        The name of the pairwise distance function to use when generating
+        pairwise distances. See the scipy ``pdist`` docs and the scikit-bio
+        functions linked under *See Also* for available metrics.
     table : biom.table.Table
         ``Table`` containing count/abundance data of observations across
         samples.
-    metric : str, optional
-        The name of the pairwise distance function to use when generating
-        pairwise distances. See the scipy ``pdist`` docs, linked under *See
-        Also*, for available metrics.
 
     Returns
     -------
@@ -91,6 +93,8 @@ def pw_distances_from_table(table, metric="braycurtis"):
 
     See Also
     --------
+    unweighted_unifrac
+    weighted_unifrac
     scipy.spatial.distance.pdist
     biom.table.Table
     pw_distances